Add JS dependency tracking via manifests.
A JS_DEPENDENCIES manifest file is added to each configuration
classpath and exported in JARs. This way, libraries may declare what
exact JS files should be included from a WebJar they depend on. (Or
what JS files they provide themselves should be included).

Example usages can be found in the build definitions for the
jasmine-test-framework and the testing example.

- This completely removes `importScripts` and fixes #457
- The project's classpath is fully available even in the post package
  stages. This is necessary to allow for discovery of dependency
  manifests and JS libraries the project depends on.