Fix #1600: Properly link before invoking Rhino.
Instead of letting Rhino re-link the classpath from scratch
every time, we link it in advance in the sbt plugin. To give
the linked code to RhinoJSEnv, we introduce a new kind of
CompleteClasspath, LinkingUnitClasspath, which is represented
by a LinkingUnit.

The instance of Linker used to do the linking is maintained
in memory with a dummy setting, so that the cache of the linker
can be leveraged.