Fix #1795: Essentials of Scala.js-defined JS classes.
This commit implements most of the specification for
Scala.js-defined JavaScript classes, as discussed in #1795.

The following features are not yet supported, and are reported
as implementation restrictions:

* Secondary constructors
* Exporting a Scala.js-defined class
* Calling the super constructor with :_* to a varargs param.

This commit does not implement the deprecations that would lead
to nicer definitions either. These aspects will come as follow
ups, since they are not critical to the general specification.