Tools refactoring
This commit substantially refactors Scala.js' linker. Notably:

Linker pipeline
---------------
The linker pipeline is divided into a frontend and a backend.

The frontend links and optimizes IR files into a `LinkingUnit`. It paves the path for linker plugins in the future.

The backend emits a linking unit into a JavaScript file. Currently there are two backends:
* BasicLinkerBackend (basic JavaScript generation, aka fastOptJS)
* ClosureBackend (optimization pass with the Google Closure compiler, aka fullOptJS)

As a side effect of this change, Rhino now runs optimized code (unless this is explicitly disabled). This further has the side effect, that the `PreLinkStage` was entirely removed.

Classpath traversal
-------------------
Classpath traversal and abstraction is fully pushed to the build system. As a consequence, the whole `org.scalajs.core.tools.classpath` package is removed. Functionally is mostly replaced by a couple of lines of sbt code.

This also fixes #1635.

IR caching
----------
In the past, each linker cached its own IR. This caused a lot of IR to be in memory multiple times. This commit replaces caching with a global IR cache. There are two main advantages:

* IR for each class is only cached once in memory.
* Linking other projects or configurations (e.g. compile v.s. test) can reuse the already cached IR.

JSEnvs
------
Since classpaths have been removed, we need a way to re-enable lazy loading in Rhino. We therefore introduce the `LinkingUnitJSEnv` that can directly accept a `LinkingUnit`. Further, we enrich the `JSEnv`s with an API to load libraries. This allows for much nicer interfaces in a couple of spots, where you essentially just pass a `JSEnv` and then run your code in it, with all dependencies already present (e.g. FrameworkDetector).

There are a couple of advantages:
- Symbol requirements are made explicit
- No duplicate code for linking in RhinoJSEnv
- Rhino can now run with optimized code (and does so by default)