Move CoreJSLibs to emitter, and make it private.
The CoreJSLibs are intimately coupled with the Emitter, and as
such, should be an implementation thereof.

It also happened to declare `jsGlobalExpr`, which is only used
by the sbt plugin, so it was moved to `ScalaJSPluginInternal`.
It is kept public there, as it might reasonably be used by
user-defined builds.