Add `@JSImport` and the generation of CommonJS modules.
This is the first and biggest step towards #2175. It contains all
the semantic changes to the language and the IR. The only thing
that is left is to actually emit ES2015 modules rather than
CommonJS modules in the emitter.

This commit contains two major changes, which are inherently linked:

* An `@JSImport` annotation which can be used on top-level native JS
  classes and objects to signify that they should be imported from
  a JavaScript module.
* A backend configuration `ModuleKind`, which instructs the emitter
  to generate a CommonJS module, rather than a script meant to be
  executed at the top-level of a JS engine.

A codebase containing a reachable `@JSImport` cannot be successfully
emitted to a top-level script, as there is no universal way of
importing things.

We could imagine emitting Node.js modules without `@JSImport` support
at first, but it would be a bit nonsensical to have modules that
cannot import from other modules.

When emitting a module, top-level `@JSExport`s become the module
exports, obviously.

Semantically, `@JSImport` and `@JSExport` are ES2015 module imports
and exports. They are therefore static, which means that a) things
cannot be conditionally imported or exported and b) their "key"
must be a compile-time string constant.

The `ModuleKind` can be specified in sbt with the new key
`scalaJSModuleKind`, which is scoped per project and per
configuration, but not per stage. Indeed, they specify the interface
with the outside JavaScript world, and that should not change in
fastOpt versus fullOpt. Obviously, setting `scalaJSModuleKind` to
`CommonJSModule` requires Node.js to be used for `run` and `test`
tasks.

The historical fact that the `scalaJSLauncher` is stage-independent
unfortunately makes `packageJSLauncher` incompatible with a
`scalaJSModuleKind` different from `NoModule`. This is because, with
modules, the contents of the launcher must reference the module ID,
which is different in fastOpt and fullOpt.

In the future, we will extend `ModuleKind` with a new kind for
actual ECMAScript 2015 modules.