Move the implemention of commonJSName to the sbt plugin.
It was previously done inside `NodeJSEnv`, which was good from
an Object-Oriented point of view. However, this required that
the API of `JSEnv` knows the notion of `commonJSName`, which is
going to be a serious issue when we isolate `jsDependencies` out
of the core.

This commit moves the implementation of the `commonJSName`
behavior in the sbt plugin, whence it can be extracted to the
`jsDependencies` plugin in the future. This has the disadvantage
that we burn `NodeJSEnv` as the only `JSEnv` for which
`commonJSName` is a thing. We can compensate for that by a
user-overridable predicate in the future `jsDependencies` plugin
if necessary.