Introduce a new surface API for JS envs with `Config` objects.
This only changes the surface API of concrete JS envs. Their
existing constructors are deprecated in favor of an overload with
a `Config` object.

This change provides in the 0.6.x series an API that can be used
in a source-compatible way between 0.6.x and 1.x. In 1.x, deeper
changes to the internal API of JS envs will be done.

We also take this opportunity to "move" `JSDOMNodeJSEnv` in a
different package `org.scalajs.jsenv.jsdomnodejs`. Since this JS
env is scheduled to be moved in a different repository in 1.x, it
should eventually be in a different package anyway.