Extract JSDOMNodeJSEnv in its own project.
So that we can move it to a different repository afterwards.

In the process, we also move the standard `NodeJSEnv` in a
dedicated project, although this one will obviously stay in the
same repository. Now that all JS environments are in different
artifacts, it makes sense that Node.js have its own artifact as
well, even though the sbt plugin will depend on that artifact.
The separation has a nice benefit that the `jsEnvsTestSuite` is
not necessary anymore, since its last tests can be moved to the
test directories of the relevant JS env implementations.