Mitigate scala-js/scala-js#3206: Send stderr of Node.js to the `JSConsole`.
This way, at least all JS environments agree on what they do with
`console.error()`, which is to merge it with `console.log()` and
send everything to the `console: JSConsole` (which is the standard
output by default).

A deeper fix would allow to separately configure stdout and stderr,
but that can only be done in Scala.js 1.x in the context of the
redesign of `JSEnv`s scala-js/scala-js#3033.