Make the constructors and underlying vals of AnyVals private.
Now that our JS artifacts do not need to be compiled with 2.10
anymore, we can truly make the underlying `val`s of `AnyVal`s
private. In the process, we also ensure that their constructors
are private.

For implicit classes, the constructor must be `private[Enclosing]`
because it is accessed by the automatically generated
`implicit def` that comes with them.

For our artifacts that still support 2.10, we also standardize on
a `val __private_foo` to make it clear that they are not part of
the API, even if the language prevents us from marking them as
actually private.