Fix scala-js/scala-js#3393: Correctly report VM crashes as failed futures.
The line in this commit was supposed to complete the future with
a failure if either `vmComplete` or `pipeResult` was failed, giving
precedence to `vmComplete`. However that line was bogus, and would
complete with a success if `pipeResult` was a success, even if
`vmComplete` was a failure.

This caused the `TestAdapter` to erroneously interpret a VM failure
as a normal exit without spawning the testing interface bridge.