Add the distinction between Script and CommonJS module in Input.
Scripts must be executed in the global scope, so that top-level
declarations end up being available to other scripts, and also as
members of the global object.

Previously, `NodeJSEnv` would only load `Input.ScriptsToLoad` as
true scripts if they were in-memory, by piping them to the standard
input of Node.js. For actual files, it used `require`, which loads
them as CommonJS modules, producing the wrong behavior for
top-level declarations.

We now introduce a separate `Input.CommonJSModulesToLoad`. For
those, `NodeJSEnv` always uses `require`, even for in-memory ones.

For `Input.ScriptsToLoad`, we use the `vm` module of Node.js and
its method `runInThisContext` in order to actually run files as
scripts, without losing source information.