Remove System.exit / __ScalaJSEnv.exitFunction
This has been added way back in the days (scala-js/scala-js#958) as a poor man's way to
support asynchronous testing.

Since then, testing has substantially matured. Notably a JSEnv is not
required anymore to be able to terminate automatically.

Therefore, there is no need anymore to support exiting in a platform
agnostic manner. Or if there is a need or such a thing, it should be
put in an external library with JS-style dynamic platform detection.

Unfortunately this makes the JSEnv test suite slightly more
complicated again (and basically reverts
6746cf3a20d2497bbc6c520666a47f7ec1cd28dd).