Remove the hack to fix percentage signs in Node.js.
Back when that hack was introduced in
1fc8d0526ba9b46049c0deefc3fe3fa93d3271b0, Node.js' `console.log`
deduplicated consecutive `%` signs in its argument, even if it was
a single argument. At the time we thought that was intentional, and
we added our hack to have it behave in a more standard way for
Scala.js.

The deduplication with only 1 argument turned out to be a bug,
recognized as such in https://github.com/nodejs/node/issues/3396,
when it was accidentally fixed in Node.js 4. It is now the
documented behavior, see
https://nodejs.org/api/util.html#util_util_format_format_args
which is referenced from
https://nodejs.org/api/console.html#console_console_log_data_args.