Import the JS Envs projects from the Scala.js core repository.
This is the initial import of the JS Envs projects
(scalajs-js-envs, scalajs-js-envs-test-kit and scalajs-env-nodejs)
from the Scala.js core repository. The history of this commit
reflects the entire history of relevant files from the core repo,
filter-branch'ed to appear as if they had always been in this repo.
This commit adds the specific setup of the build and tests.